requirements.txt
geopandas==0.14.4
graphlearning==1.5.1
infomap==2.7.1
matplotlib==3.8.2
netlsd==1.0.2
networkx==3.2.1
numpy==1.26.4
pandas==2.2.2
pybind11==2.12.0
PyGSP==0.5.1
python_example==0.0.1
python_igraph==0.11.5
rustworkx==0.14.2
scikit_learn==1.3.2
scipy==1.13.0
setuptools==58.1.0
tabulate==0.9.0
toml==0.10.2
ucimlrepo==0.0.6
